Renew the entire system.
The main features of this renewal are abolition of package prefixes
and support for "embedded" packages.

Embedded packages are those each of which is not only one existing
in the building environment but bundled with its container package,
and referenced with its name by foreign targets, whose definitions are
preceded by loading of the container package.