Remove unused code for legacy package variable
Legacy package variable was intially introduced at 6c86401b and
an utility to set variable was added at f4998206, but we switched to
currently working package variable at ebaf32e5 and the legacy one was
not working from that time. Since all commits between 6c86401b and ebaf32e5
were before the release of version 0.0.4, we got to be able to assume
that there are no projects using legacy package variables.